Refresh Anthropic Gemini and Groq model metadata
