+++ src/core/cost/calculator.rs
@@ -102,15 +102,26 @@ pub fn generic_cost_per_token(
 
 /// Get model pricing information
 pub fn get_model_pricing(model: &str, provider: &str) -> Result<ModelPricing, CostError> {
-    match provider.to_lowercase().as_str() {
+    let normalized_provider = crate::core::pricing::normalize_pricing_provider(provider);
+
+    match normalized_provider.as_str() {
         "openai" => get_pricing_with_shared_source(model, &["openai"], get_openai_pricing),
         "anthropic" => get_pricing_with_shared_source(model, &["anthropic"], get_anthropic_pricing),
         "azure" => get_azure_pricing(model),
-        "vertex_ai" | "vertexai" => {
+        "vertex_ai" => {
             get_pricing_with_shared_source(model, &["vertex_ai", "google"], get_vertex_ai_pricing)
         }
+        "gemini" => {
+            get_pricing_with_shared_source(model, &["gemini", "vertex_ai"], get_vertex_ai_pricing)
+        }
+        "groq" => get_pricing_with_shared_source(model, &["groq"], |model| {
+            Err(CostError::ModelNotSupported {
+                model: model.to_string(),
+                provider: "groq".to_string(),
+            })
+        }),
         "deepseek" => get_pricing_with_shared_source(model, &["deepseek"], get_deepseek_pricing),
-        "xiaomi_mimo" | "xiaomi" | "mimo" => {
+        "xiaomi_mimo" => {
             get_pricing_with_shared_source(model, &["xiaomi_mimo", "xiaomi", "mimo"], |model| {
                 Err(CostError::ModelNotSupported {
                     model: model.to_string(),
@@ -120,7 +131,7 @@ pub fn get_model_pricing(model: &str, provider: &str) -> Result<ModelPricing, Co
         }
         "moonshot" => get_pricing_with_shared_source(model, &["moonshot"], get_moonshot_pricing),
         "minimax" => get_pricing_with_shared_source(model, &["minimax"], get_minimax_pricing),
-        "zhipu" | "zhipuai" | "glm" | "zai" => {
+        "zhipuai" => {
             get_pricing_with_shared_source(model, &["zhipuai", "glm", "zai"], get_zhipu_pricing)
         }
         _ => Err(CostError::ProviderNotSupported {
@@ -166,8 +177,7 @@ fn get_shared_model_pricing(
 
         for model_id in candidates {
             let model_id_lower = model_id.to_lowercase();
-            let matches = model_id_lower == model_lower || model_id_lower.contains(&model_lower);
-            if matches
+            if is_shared_model_match(&model_id_lower, &model_lower)
                 && let Some(info) = db.get_model_info(&model_id)
                 && litellm_provider_matches(&info.litellm_provider, provider_aliases)
             {
@@ -179,20 +189,40 @@ fn get_shared_model_pricing(
     None
 }
 
-fn litellm_provider_matches(provider: &str, aliases: &[&str]) -> bool {
-    let provider = normalize_pricing_provider(provider);
-    aliases
-        .iter()
-        .any(|alias| normalize_pricing_provider(alias) == provider)
+fn is_shared_model_match(candidate: &str, requested: &str) -> bool {
+    fn model_id_matches(candidate: &str, requested: &str) -> bool {
+        if candidate == requested {
+            return true;
+        }
+
+        candidate
+            .strip_prefix(requested)
+            .and_then(|suffix| suffix.strip_prefix('-'))
+            .is_some_and(|suffix| {
+                let digit_prefix_len = suffix.chars().take_while(|ch| ch.is_ascii_digit()).count();
+                digit_prefix_len >= 4
+                    && suffix
+                        .chars()
+                        .all(|ch| ch.is_ascii_alphanumeric() || ch == '-' || ch == '_')
+            })
+    }
+
+    if candidate == requested {
+        return true;
+    }
+
+    model_id_matches(candidate, requested)
+        || candidate
+            .rsplit_once('/')
+            .map(|(_, model_id)| model_id_matches(model_id, requested))
+            .unwrap_or(false)
 }
 
-fn normalize_pricing_provider(provider: &str) -> String {
-    match provider.to_lowercase().replace('-', "_").as_str() {
-        "vertexai" | "google" => "vertex_ai".to_string(),
-        "zhipu" | "glm" | "zai" => "zhipuai".to_string(),
-        "xiaomi" | "mimo" => "xiaomi_mimo".to_string(),
-        other => other.to_string(),
-    }
+fn litellm_provider_matches(provider: &str, aliases: &[&str]) -> bool {
+    let provider = crate::core::pricing::normalize_pricing_provider(provider);
+    aliases
+        .iter()
+        .any(|alias| crate::core::pricing::normalize_pricing_provider(alias) == provider)
 }
 
 fn litellm_to_cost_pricing(
