+++ src/core/cost/utils.rs
@@ -200,6 +200,7 @@ pub fn get_model_category(model: &str) -> &'static str {
         || model_lower.contains("gpt-5-codex")
         || model_lower.contains("o3-pro")
         || model_lower.contains("gpt-4o")
+        || model_lower.contains("claude-opus-4-8")
         || model_lower.contains("claude-opus-4-7")
         || model_lower.contains("claude-opus-4-6")
         || model_lower.contains("claude-opus-4-5")
