+++ src/core/pricing/tests.rs
@@ -161,6 +161,98 @@ fn xiaomi_mimo_provider_aliases_share_pricing_rows() {
 }
 
 #[test]
+fn extended_pricing_has_current_anthropic_gemini_and_groq_rates() {
+    let Ok(db) = PricingDatabase::from_default_source() else {
+        panic!("shared pricing source should load");
+    };
+    let usage = Usage::new(1000, 500);
+
+    assert!(
+        (db.calculate_for_provider("anthropic", "claude-opus-4-8", &usage) - 0.0175).abs() < 1e-12
+    );
+    assert!(
+        (db.calculate_for_provider("anthropic", "claude-sonnet-4-6", &usage) - 0.0105).abs()
+            < 1e-12
+    );
+    assert!(
+        (db.calculate_for_provider("anthropic", "claude-haiku-4-5-20251001", &usage) - 0.0035)
+            .abs()
+            < 1e-12
+    );
+
+    assert!(
+        (db.calculate_for_provider("vertex_ai", "gemini-3.1-flash-lite", &usage) - 0.001).abs()
+            < 1e-12
+    );
+    assert!(
+        (db.calculate_for_provider("gemini", "gemini-3.1-flash-lite", &usage) - 0.001).abs()
+            < 1e-12
+    );
+    assert!(
+        db.get_provider_models("gemini")
+            .contains(&"gemini/gemini-2.5-flash".to_string()),
+        "gemini provider should list prefixed Gemini 2.5 rows"
+    );
+    assert_eq!(
+        db.calculate_for_provider("vertex_ai", "gemini-3.1-flash", &usage),
+        0.0,
+        "Gemini Flash must not borrow Flash-Lite pricing"
+    );
+
+    let long_usage = Usage::new(300_000, 1_000);
+    let long_context_cost = 300_000.0 * 0.000004 + 1_000.0 * 0.000018;
+    assert!(
+        (db.calculate_for_provider("vertex_ai", "gemini-3.1-pro-preview", &long_usage)
+            - long_context_cost)
+            .abs()
+            < 1e-12
+    );
+
+    assert!(
+        (db.calculate_for_provider("groq", "llama-3.3-70b-versatile", &usage) - 0.000985).abs()
+            < 1e-12
+    );
+    assert!(
+        (db.calculate_for_provider("groq", "openai/gpt-oss-120b", &usage) - 0.00045).abs() < 1e-12
+    );
+    assert!((db.calculate_for_provider("groq", "gpt-oss-120b", &usage) - 0.00045).abs() < 1e-12);
+
+    let Some(whisper) = db.get_model_info("whisper-large-v3-turbo") else {
+        panic!("Groq Whisper Turbo ASR pricing entry should exist");
+    };
+    assert_eq!(whisper.cost_per_second, Some(0.000011111111111111112));
+    assert_eq!(whisper.mode, "audio_transcription");
+}
+
+#[test]
+fn provider_aliases_share_normalized_pricing_rows() {
+    let Ok(db) = PricingDatabase::from_default_source() else {
+        panic!("shared pricing source should load");
+    };
+    let usage = Usage::new(1000, 500);
+
+    assert_eq!(
+        db.calculate_for_provider("google", "gemini-1.5-flash", &usage),
+        db.calculate_for_provider("vertex_ai", "gemini-1.5-flash", &usage)
+    );
+    assert!(
+        db.get_provider_models("google")
+            .contains(&"gemini-1.5-flash".to_string()),
+        "google should list Vertex AI pricing rows"
+    );
+
+    assert_eq!(
+        db.calculate_for_provider("zhipuai", "glm-5", &usage),
+        db.calculate_for_provider("zhipu", "glm-5", &usage)
+    );
+    assert!(
+        db.get_provider_models("zhipuai")
+            .contains(&"glm-5".to_string()),
+        "zhipuai should list Zhipu pricing rows"
+    );
+}
+
+#[test]
 fn test_model_info() {
     let db = PricingDatabase::default();
