+++ src/core/providers/anthropic/mod.rs
@@ -93,7 +93,7 @@ pub fn validate_api_key(api_key: &str) -> Result<(), String> {
 
 /// Default
 pub fn default_model() -> &'static str {
-    "claude-opus-4-7"
+    "claude-opus-4-8"
 }
 
 /// Model
@@ -199,6 +199,7 @@ mod tests {
 
     #[test]
     fn test_model_support() {
+        assert!(is_model_supported("claude-opus-4-8"));
         assert!(is_model_supported("claude-opus-4-7"));
         assert!(is_model_supported("claude-3-haiku-20240307"));
         assert!(!is_model_supported("gpt-4"));
@@ -206,7 +207,7 @@ mod tests {
 
     #[test]
     fn test_model_features() {
-        let features = get_model_features("claude-opus-4-7");
+        let features = get_model_features("claude-opus-4-8");
         assert!(features.is_some());
 
         let features = features.unwrap();
@@ -218,7 +219,7 @@ mod tests {
     #[test]
     fn test_feature_support() {
         assert!(model_supports_feature(
-            "claude-opus-4-7",
+            "claude-opus-4-8",
             ModelFeature::ComputerUse
         ));
         assert!(!model_supports_feature(
@@ -233,8 +234,8 @@ mod tests {
 
     #[test]
     fn test_cost_estimation() {
-        let cost = estimate_cost("claude-opus-4-7", 1000, 500);
-        let cost = cost.expect("claude-opus-4-7 should have Anthropic pricing");
+        let cost = estimate_cost("claude-opus-4-8", 1000, 500);
+        let cost = cost.expect("claude-opus-4-8 should have Anthropic pricing");
         assert!((cost - 0.0175).abs() < f64::EPSILON);
     }
 
@@ -242,13 +243,14 @@ mod tests {
     fn test_supported_models_list() {
         let models = supported_models();
         assert!(!models.is_empty());
+        assert!(models.contains(&"claude-opus-4-8".to_string()));
         assert!(models.contains(&"claude-opus-4-7".to_string()));
         assert!(models.contains(&"claude-3-haiku-20240307".to_string()));
     }
 
     #[test]
     fn test_default_model() {
-        assert_eq!(default_model(), "claude-opus-4-7");
+        assert_eq!(default_model(), "claude-opus-4-8");
         assert!(is_model_supported(default_model()));
     }
