+++ src/core/providers/gemini/models/mod.rs
@@ -561,7 +561,7 @@ mod tests {
             "gemini-3-flash-preview",
             "gemini-3.1-pro-preview",
             "gemini-3.1-flash",
-            "gemini-3.1-flash-lite-preview",
+            "gemini-3.1-flash-lite",
             "gemini-2.0-flash-exp",
             "gemini-1.5-pro",
             "gemini-1.5-flash",
