+++ src/sdk/config.rs
@@ -185,12 +185,12 @@ impl SdkConfigBuilder {
             api_key: api_key.to_string(),
             base_url: None,
             models: vec![
+                "claude-opus-4-8".to_string(),
                 "claude-opus-4-7".to_string(),
                 "claude-sonnet-4-6".to_string(),
                 "claude-haiku-4-5-20251001".to_string(),
-                "claude-3-5-haiku-20241022".to_string(),
                 "claude-haiku-4-5".to_string(),
-                "claude-haiku-4-5-20251001".to_string(),
+                "claude-3-5-haiku-20241022".to_string(),
             ],
             enabled: true,
             weight: 1.0,
@@ -611,7 +611,7 @@ mod tests {
             config.providers[0]
                 .models
                 .iter()
-                .any(|m| m == "claude-3-5-haiku-20241022")
+                .any(|m| m == "claude-opus-4-8")
         );
         assert!(
             config.providers[0]
